pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-itau"
version = "0.1.0"
description = "Itaú Uruguay banking — CLI, REST API, and MCP server"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "fastmcp>=2.0.0",
    "keyring>=25.0.0",
]

[project.scripts]
my-itau = "my_itau.cli:main"

[tool.setuptools.packages.find]
include = ["my_itau*"]
